:fire: [#10] REMOVE valuerenderers and put it in one file
